+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Bump constraint on python to >=3.8,<3.12 (upper constraint is required by pex)
 - Update dependencies
+- [aws] Improve `exec` error messages
 - [jumpcloud,okta] Verify SSO URLs before making SAML requests
 - [jumpcloud,okta] Verify that plugin is configured before making SAML requests